Move parser out of core modules
