+++ src/core/mod.rs
@@ -1,3 +1,2 @@
 pub mod central_disk_logger;
-pub mod parser;
 pub mod thread_manager;
