+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod ext;
 pub mod gui;
 pub mod ingestor;
 pub mod logging;
+pub mod parser;
 pub mod pb;
 pub mod pipeline;
